Remove obsolete scripts/generatecss.php
Clearly broken, as it tries to include a non-existing Resources.php file.

Also it's pretending to be MediaWiki 1.23, which is plenty old.

It was supposed to be run via the Makefile, but that's been broken since at
least 111b6714447. If this functionality of identifying duplicate CSS rules
is still wanted, I'm sure there's a better way to do it than with a hacky PHP
script and a ruby tool.

Mainly removed since it's causing a phan-taint-check-plugin warning (T202383).

Change-Id: Id1d117f41f93adec4bba7191843aa415c4049d5f