Merge 'ext.echo.controller' module into 'ext.echo.ui'
This is not used anywhere else and doesn't appear to be something
that is intended to be usable or loadable as its own entry point.

The name 'ext.echo.ui' could be improved to better reflect the
bundle purpose it represents, but thats a breaking change for a
separate commit at the discretion of the maintainers' preferences.

Change-Id: If9cf287917a51de96a36d534b7c0cfadf1accf9a