Use {{formatnum:}} for 99+ in default echo-badge-count message
As it stands currently, every language that uses different numerals than
English does has to override the echo-badge-count message to provide
their own version of "99+". If a language doesn't do that, numbers up to
99 will display using the correct numerals, but "99+" will be displayed
verbatim. Some affected languages have overridden this message, but most
haven't.

Instead, use "{{formatnum:99}}+" in the English message. This is a no-op
in English and other languages that use the same numerals as English,
but automatically converts the number 99 in languages that use different
numerals. This way the only languages that need to override the message
are ones that use a different symbol than "+" for expressing "or more".

Change-Id: I247260ed9bbb40cff88e5ab873072269221b0ad4