Hide "system" categories in notification preferences
Make sure that the system-noemail and system-emailonly categories are
hidden, the way the system category is, by setting their no-dismiss
setting to "all". This should be unnecessary, because e.g.
system-noemail doesn't allow email, so just setting no-dismiss to "web"
should suffice, but
EchoAttributeManager::isCategoryDisplayedInPreferences() isn't smart
enough to see that. Doing it this way is also more future-proof for if
we ever add more notification methods in addition to web and email.

Change-Id: I00641d43a4eb1d3e2d3ea285682cf900733b1855